Change run records to behavior subject.
